+++ Panels/Panels.js
@@ -220,7 +220,7 @@ const PanelsBase = kind({
 	handlers: {
 		onBack: handle(
 			adaptEvent(
-				(ev, {index}) => ({index: Math.max(index - 1, 0)}),
+				(ev, {index}) => ({type: 'onBack', index: Math.max(index - 1, 0)}),
 				forwardWithPrevent('onBack')
 			)
 		)